% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicInitParameters.R
\name{basic_Init_Parameters}
\alias{basic_Init_Parameters}
\title{Initialize the parameters in a basic way (zero for the conditional mean,
uniform for weights, identity for covariance matrices, and uniformly distributed forthe clustering)}
\usage{
basic_Init_Parameters(n, p, m, k)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{m}{size of the response}

\item{k}{number of clusters}
}
\value{
list with phiInit, rhoInit,piInit,gamInit
}
\description{
Initialize the parameters in a basic way (zero for the conditional mean,
uniform for weights, identity for covariance matrices, and uniformly distributed forthe clustering)
}

