% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSampleInputs.R
\name{basicInitParameters}
\alias{basicInitParameters}
\title{Initialize the parameters in a basic way (zero for the conditional mean, uniform for weights,
identity for covariance matrices, and uniformly distributed for the clustering)}
\usage{
basicInitParameters(n, p, m, k)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{m}{size of the response}

\item{k}{number of clusters}
}
\value{
list with phiInit, rhoInit,piInit,gamInit
}
\description{
Initialize the parameters in a basic way (zero for the conditional mean, uniform for weights,
identity for covariance matrices, and uniformly distributed for the clustering)
}

