% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGridLambda.R
\name{computeGridLambda}
\alias{computeGridLambda}
\title{computeGridLambda}
\usage{
computeGridLambda(
  phiInit,
  rhoInit,
  piInit,
  gamInit,
  X,
  Y,
  gamma,
  mini,
  maxi,
  eps,
  fast
)
}
\arguments{
\item{phiInit}{value for phi}

\item{rhoInit}{for rho}

\item{piInit}{for pi}

\item{gamInit}{value for gamma}

\item{X}{matrix of covariates (of size n*p)}

\item{Y}{matrix of responses (of size n*m)}

\item{gamma}{power of weights in the penalty}

\item{mini}{minimum number of iterations in EM algorithm}

\item{maxi}{maximum number of iterations in EM algorithm}

\item{eps}{threshold to stop EM algorithm}

\item{fast}{boolean to enable or not the C function call}
}
\value{
the grid of regularization parameters for the Lasso estimator. The output is a vector with nonnegative values that are relevant
to be considered as regularization parameter as they are equivalent to a 0 in the regression parameter.
}
\description{
Construct the data-driven grid for the regularization parameters used for the Lasso estimator
}
