% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardSimilarModels.R
\name{discardSimilarModels}
\alias{discardSimilarModels}
\title{Discard models which have the same relevant variables}
\usage{
discardSimilarModels(B1, B2, glambda, rho, pi)
}
\arguments{
\item{B1}{array of relevant coefficients (of size p*m*length(gridlambda))}

\item{B2}{array of irrelevant coefficients (of size p*m*length(gridlambda))}

\item{glambda}{grid of regularization parameters (vector)}

\item{rho}{covariance matrix (of size m*m*K*size(gridLambda))}

\item{pi}{weight parameters (of size K*size(gridLambda))}
}
\value{
a list with update B1, B2, glambda, rho and pi, and ind the vector of indices
of selected models.
}
\description{
Discard models which have the same relevant variables
}

