% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardSimilarModels2.R
\name{discardSimilarModels2}
\alias{discardSimilarModels2}
\title{Similar to discardSimilarModels, for Lasso-rank procedure (focus on columns)}
\usage{
discardSimilarModels2(B1, rho, pi)
}
\arguments{
\item{B1}{array of relevant coefficients (of size p*m*length(gridlambda))}

\item{rho}{covariance matrix}

\item{pi}{weight parameters}
}
\value{
a list with B1, in, rho, pi
}
\description{
Similar to discardSimilarModels, for Lasso-rank procedure (focus on columns)
}

