% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardSimilarModels.R
\name{discardSimilarModels_EMGrank}
\alias{discardSimilarModels_EMGrank}
\title{Discard models which have the same relevant variables
  - for Lasso-rank procedure (focus on columns)}
\usage{
discardSimilarModels_EMGrank(B1, rho, pi)
}
\arguments{
\item{B1}{array of relevant coefficients (of size p*m*length(gridlambda))}

\item{rho}{covariance matrix}

\item{pi}{weight parameters}
}
\value{
a list with B1, in, rho, pi
}
\description{
Discard models which have the same relevant variables
  - for Lasso-rank procedure (focus on columns)
}

