% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIO.R
\name{generateIO}
\alias{generateIO}
\title{Generate a sample of (X,Y) of size n}
\usage{
generateIO(covX, covY, pi, beta, n)
}
\arguments{
\item{covX}{covariance for covariates (of size p*p*K)}

\item{covY}{covariance for the response vector (of size m*m*K)}

\item{pi}{proportion for each cluster}

\item{beta}{regression matrix}

\item{n}{sample size}
}
\value{
list with X and Y
}
\description{
Generate a sample of (X,Y) of size n
}

