% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSampleInputs.R
\name{generateXYdefault}
\alias{generateXYdefault}
\title{Generate a sample of (X,Y) of size n with default values}
\usage{
generateXYdefault(n, p, m, k)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{m}{size of the response}

\item{k}{number of clusters}
}
\value{
list with X and Y
}
\description{
Generate a sample of (X,Y) of size n with default values
}

