% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicesSelection.R
\name{indicesSelection}
\alias{indicesSelection}
\title{Construct the set of relevant indices}
\usage{
indicesSelection(phi, thresh = 1e-06)
}
\arguments{
\item{phi}{regression matrix, of size p*m}

\item{thresh}{threshold to say a cofficient is equal to zero}
}
\value{
a list with A, a matrix with relevant indices (size = p*m) and B, a 
         matrix with irrelevant indices (size = p*m)
}
\description{
Construct the set of relevant indices
}

