% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{modelSelection}
\alias{modelSelection}
\title{Among a collection of models, this function constructs a subcollection of models with
models having strictly different dimensions, keeping the model which minimizes
the likelihood if there were several with the same dimension}
\usage{
modelSelection(LLF)
}
\arguments{
\item{LLF}{a matrix, the first column corresponds to likelihoods for several models
the second column corresponds to the dimensions of the corresponding models.}
}
\value{
a list with indices, a vector of indices selected models,
			 and D1, a vector of corresponding dimensions
}
\description{
Among a collection of models, this function constructs a subcollection of models with
models having strictly different dimensions, keeping the model which minimizes
the likelihood if there were several with the same dimension
}

