% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_valse.R
\name{plot_valse}
\alias{plot_valse}
\title{Plot}
\usage{
plot_valse(X, Y, model, comp = FALSE, k1 = NA, k2 = NA)
}
\arguments{
\item{X}{matrix of covariates (of size n*p)}

\item{Y}{matrix of responses (of size n*m)}

\item{model}{the model constructed by valse procedure}

\item{comp}{TRUE to enable pairwise clusters comparison}

\item{k1}{index of the first cluster to be compared}

\item{k2}{index of the second cluster to be compared}
}
\value{
No return value (only plotting).
}
\description{
A function which plots relevant parameters.
}
