\name{valse-package}
\alias{valse-package}
\alias{valse}
\docType{package}

\title{
	\packageTitle{valse}
}

\description{
	\packageDescription{valse}
}

\details{
	The package devtools should be useful in development stage, since we rely on testthat for
	unit tests, and roxygen2 for documentation. knitr is used to generate the package vignette.
	Concerning the other suggested packages:
	\itemize{
		\item{parallel (generally) permits to run the bootstrap method faster.}
	}

	The three main functions are ...
}

\author{
	\packageAuthor{valse}

	Maintainer: \packageMaintainer{valse}
}

%\references{
%	TODO: Literature or other references for background information
%}

%\examples{
%	TODO: simple examples of the most important functions
%}
