% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valse.R
\name{valse}
\alias{valse}
\title{Main function}
\usage{
valse(X, Y, procedure = "LassoMLE", selecMod = "BIC", gamma = 1,
  mini = 10, maxi = 100, eps = 1e-04, kmin = 2, kmax = 5,
  rang.min = 1, rang.max = 10)
}
\arguments{
\item{X}{matrix of covariates (of size n*p)}

\item{Y}{matrix of responses (of size n*m)}

\item{procedure}{among 'LassoMLE' or 'LassoRank'}

\item{selecMod}{method to select a model among 'SlopeHeuristic', 'BIC', 'AIC'}

\item{gamma}{integer for the power in the penaly, by default = 1}

\item{mini}{integer, minimum number of iterations in the EM algorithm, by default = 10}

\item{maxi}{integer, maximum number of iterations in the EM algorithm, by default = 100}

\item{eps}{real, threshold to say the EM algorithm converges, by default = 1e-4}

\item{kmin}{integer, minimum number of clusters, by default = 2}

\item{kmax}{integer, maximum number of clusters, by default = 10}

\item{rang.min}{integer, minimum rank in the low rank procedure, by default = 1}

\item{rang.max}{integer, maximum rank in the}
}
\value{
a list with estimators of parameters
}
\description{
Main function
}

