% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_bin.R
\name{vec_bin}
\alias{vec_bin}
\title{A function needed in initSmallEM}
\usage{
vec_bin(X, r)
}
\arguments{
\item{X}{vector with integer values}

\item{r}{integer}
}
\value{
a list with Z (a binary vector of size the size of X) and indices where Z is equal to 1
}
\description{
A function needed in initSmallEM
}

