% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateXY.R
\name{generateXY}
\alias{generateXY}
\title{generateXY}
\usage{
generateXY(n, prop, meanX, beta, covX, covY)
}
\arguments{
\item{n}{sample size}

\item{prop}{proportion for each cluster}

\item{meanX}{matrix of group means for covariates (of size p)}

\item{beta}{regression matrix, of size p*m*k}

\item{covX}{covariance for covariates (of size p*p)}

\item{covY}{covariance for the response vector (of size m*m)}
}
\value{
list with X (of size n*p) and Y (of size n*m)
}
\description{
Generate a sample of (X,Y) of size n
}
