% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridLambda.R
\name{gridLambda}
\alias{gridLambda}
\title{Construct the data-driven grid for the regularization parameters used for the Lasso estimator}
\usage{
gridLambda(phiInit, rhoInit, piInit, gamInit, X, Y, gamma, mini, maxi, tau)
}
\arguments{
\item{phiInit}{value for phi}

\item{piInit}{value for pi}

\item{gamInit}{value for gamma}

\item{mini}{minimum number of iterations in EM algorithm}

\item{maxi}{maximum number of iterations in EM algorithm}

\item{tau}{threshold to stop EM algorithm}

\item{rhoInt}{value for rho}
}
\value{
the grid of regularization parameters
}
\description{
Construct the data-driven grid for the regularization parameters used for the Lasso estimator
}

