% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSmallEM.R
\name{initSmallEM}
\alias{initSmallEM}
\title{initSmallEM}
\usage{
initSmallEM(k, X, Y, fast)
}
\arguments{
\item{k}{number of components}

\item{X}{matrix of covariates (of size n*p)}

\item{Y}{matrix of responses (of size n*m)}

\item{fast}{boolean to enable or not the C function call}
}
\value{
a list with phiInit (the regression parameter reparametrized),
rhoInit (the covariance parameter reparametrized), piInit (the proportion parameter is the
mixture model), gamInit (the conditional expectation)
}
\description{
initialization of the EM algorithm
}
