% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSmallEM.R
\name{initSmallEM}
\alias{initSmallEM}
\title{initialization of the EM algorithm}
\usage{
initSmallEM(k, X, Y)
}
\arguments{
\item{k}{number of components}

\item{X}{matrix of covariates (of size n*p)}

\item{Y}{matrix of responses (of size n*m)}
}
\value{
a list with phiInit, rhoInit, piInit, gamInit
}
\description{
initialization of the EM algorithm
}

