% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVariables.R
\name{selectVariables}
\alias{selectVariables}
\title{selectVariables}
\usage{
selectVariables(phiInit, rhoInit, piInit, gamInit, mini, maxi, gamma, glambda,
  X, Y, thresh, tau, ncores = 1)
}
\arguments{
\item{phiInit}{an initial estimator for phi (size: p*m*k)}

\item{rhoInit}{an initial estimator for rho (size: m*m*k)}

\item{piInit}{an initial estimator for pi (size : k)}

\item{gamInit}{an initial estimator for gamma}

\item{mini}{minimum number of iterations in EM algorithm}

\item{maxi}{maximum number of iterations in EM algorithm}

\item{gamma}{power in the penalty}

\item{glambda}{grid of regularization parameters}

\item{X}{matrix of regressors}

\item{Y}{matrix of responses}

\item{tau}{threshold to say that EM algorithm has converged}

\item{thres}{threshold to consider a coefficient to be equal to 0}
}
\value{
a list of outputs, for each lambda in grid: selected,Rho,Pi
}
\description{
It is a function which construct, for a given lambda, the sets of relevant variables.
}
\examples{
TODO

}

