\name{valse-package}
\alias{valse-package}
\alias{valse}
\docType{package}

\title{
	\packageTitle{valse}
}

\description{
	\packageDescription{valse}
}

\details{
  Two methods are implemented to cluster data with finite mixture
  regression models. Those procedures deal with high-dimensional covariates and
  responses through a variable selection procedure based on the Lasso estimator.
  
  The main function is runValse(), which calls all other functions.
  See also plot_valse() which plots the relevant parameters after a run.
}

\author{
	\packageAuthor{valse}

	Maintainer: \packageMaintainer{valse}
}

%\references{
%	TODO: Literature or other references for background information
%}

%\examples{
%	TODO: simple examples of the most important functions
%}
